==> 04_secretome/notebook/03_fold_secretome.Rmd <==
---
title: "03_fold_secretome.md"
author: "Gulnara Tagridzhanova"
date: "12/01/2024"
output: 
  html_document:
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "/Users/gol22pin/Documents/Lichen_project_TSL/10_lichen_effectors/notebook")

```
* **Summary:** do CollabFold predictions on all secreted proteins

## 1. Prepared the sequences
* Took the fasta of the proteins that are secreted according to all three tools from `analysis_and_temp_files/01_predicting_effectors/all_secreted.fa`
* Split it into individual files (one sequence per file)
```{r, eval=FALSE}
library(Biostrings)
secreted_fa<-readAAStringSet("../analysis_and_temp_files/01_predicting_effectors/all_secreted.fa")

for (i in 1:length(secreted_fa)) {
    header <- names(secreted_fa)[i]
    sequence <- as.character(secreted_fa[i])
    
    # Combine header and sequence
    combined_data <- paste0(">", header, "\n",sequence)
    
    # Save the combined data as a separate FASTA file
    file_name <- paste0("../analysis_and_temp_files/03_fold_secretome/individual_fastas/",header, ".fasta")
    writeLines(combined_data, file_name)
    
}

```

* Took first 100 of sequences as a subset to try. Put them into `analysis_and_temp_files/03_fold_secretome/individual_fastas_subset/`
```{r, eval=FALSE}
library(Biostrings)
secreted_fa<-readAAStringSet("../analysis_and_temp_files/01_predicting_effectors/all_secreted.fa")

for (i in 1:100) {
    header <- names(secreted_fa)[i]
    sequence <- as.character(secreted_fa[i])
    
    # Combine header and sequence
    combined_data <- paste0(">", header, "\n",sequence)
    
    # Save the combined data as a separate FASTA file
    file_name <- paste0("../analysis_and_temp_files/03_fold_secretome/individual_fastas_subset/",header, ".fasta")
    writeLines(combined_data, file_name)
    
}


```

## 2. ColabFold structure predictions
* Ran on first 100 proteins. 670.01 compute units were available at the start, and 542.13 after. 
  * This corresponds to 1.28 unit per protein. 
  * All 608 proteins should require 777.5 units 
  * At the rate 500 units for 50 pounds, this should cost 77.5 pounds
* Ran on the remaining proteins

#### Control: compared to AlphaFold predictions
* Took one model, FUN_004252-T1 and compared predictions from ColabFold and from AlphaFold from the hpc
* The match is great. All ordered parts are aligned exactly, disorded parts are shifted but still look similar

#### File locations 
* Complete zipped outputs of ColabFold: `analysis_and_temp_files/03_fold_secretome/colab_fold_complete_output`
* Best models fro each 608 proteins: `analysis_and_temp_files/03_fold_secretome/colab_fold_best_model`
* Best models for the 91 putative effectors, separated on Adam's request: `analysis_and_temp_files/03_fold_secretome/colab_fold_only_putative_effectors`

```{r, eval=F}
while read -r line; do for file in $line; do cp analysis_and_temp_files/03_fold_secretome/colab_fold_best_model/"$file"* analysis_and_temp_files/03_fold_secretome/colab_fold_only_putative_effectors; done; done < analysis_and_temp_files/01_predicting_effectors/consensus_list.txt 
```
